requirements.txt
torch
torchvision
tqdm
timm
Pillow
flask